requirements.txt
Flask
pymongo[srv]
python-dotenv
Flask-Cors
web3 